separate term choose
